Install java-11-openjdk instead of java-latest-openjdk
Java 8, 11 and 17 are the LTS versions. Currently, most Linux distros
mainly use Java 11, but still provide Java 8.

java-latest-openjdk is the shortname for the latest Java version,
which may be unstable (e.g., Java 18). Thus, it's not a good idea
to install java-latest-openjdk, because some software may not work
with latest java versions (e.g., TauP).

Note that, the changes only affect Fedora, not Ubuntu. On Ubunut,
default-jdk always points to the latest stable Java version.